requirements.txt
DJANGO
psycopg2-binary
djangorestframework
